pyproject.toml
[project]
name = "image-caption"
version = "0.1.0"
description = "Industrial-grade image captioning application"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "pillow>=12.0.0",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.0.1",
    "gunicorn>=21.2.0",
    "Flask-Limiter>=3.7.0",
    "Flask-WTF>=1.2.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"


[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ipykernel>=6.29.0",
    "ruff>=0.3.0",
]
